requirements.txt
disnake==2.9.0
loguru==0.7.2
python-dotenv==1.0.0
Requests==2.31.0
